requirements.txt
pandas>=2.2
pyarrow>=16.0
python-dotenv>=1.0
pydantic>=2.7
tqdm>=4.66
rapidfuzz>=3.9
sqlalchemy>=2.0
# Data sources
soccerdata>=2.8
fantraxapi>=3.2
# API
fastapi>=0.115
uvicorn[standard]>=0.30
